music/compare-and-contrast: add tool for diffing changes to ID3 data.
When you start the script, it extracts the ID3 frame data using
`id3dump -f', and prompts you to perform some operation that will make
modifications to the data (such as updating the MP3 tags with
MusicBrainz Picard).  When the change is made, you can press "enter"
to re-extract the frame data and compare with the previous state.