NCFO: Prepare download scripts for Kata's opera debut. =)
We now have two versions of the download scripts (one for opera, one
for the Science Festival chorus), because the new architecture is
pretty different than the old one; we need to deal with things like
table parsing to get the lists correct, and don't need things like
manual track reordering.  I'll try to reuse some of the new stuff for
Science Festival chorus (e.g. use process-files as backend, with
current copy-files jus generating a list as input)...  Also, the
Drupal page IDs don't change (for the chorus at least), so next year's
chorus stuff wants the same URLs as last chorus but not opera.