[NCFO] extras2process: read directives and only add matching extras.
In addition to the list of extras directories on the command line, the
script now also reads in a list of already prepared processing
directives on stdin.  The output is all of the directives seen on
stdin, *plus* all of the corresponding extras.  Extras for directories
not seen on stdin will be ignored, with a warning message.

This means this script must now be invoked as a "filter" for an
existing process stream, not just added to it at the end.