provision-encrypted-backup: Fix issues
The align-partition script has changed names.

Also, add a hackish delay to wait_for().  Sometimes the deletion and
creation of a partition are delayed slightly by the kernel, causing
wait_for to succeed but the operation that follows it to fail.  The
right fix would be to add a "wait_for_gone" as an inversee of
wait_for, and call it after partition deletion, but I can't be
bothered.